+++ cachetclient/v1/__init__.py
@@ -3,9 +3,9 @@ from cachetclient.v1.client import Client  # noqa
 from cachetclient.v1.subscribers import Subscriber  # noqa
 from cachetclient.v1.components import Component  # noqa
 from cachetclient.v1.component_groups import ComponentGroup  # noqa
+from cachetclient.v1.incidents import Incident  # noqa
 from cachetclient.v1.incident_updates import IncidentUpdate  # noqa
 from cachetclient.v1.metrics import Metric  # noqa
 from cachetclient.v1.metric_points import MetricPoint  # noqa
+from cachetclient.v1.schedules import Schedule  # noqa
 from cachetclient.v1 import enums  # noqa
-
-__version__ = '1.1.0'