+++ cachetclient/v1/incidents.py
@@ -74,6 +74,15 @@ class Incident(Resource):
         self._data['visible'] = value
 
     @property
+    def stickied(self) -> int:
+        """bool: Get or set sticky value of the incident (cachet 2.4)"""
+        return self.get('stickied') == 1
+
+    @stickied.setter
+    def stickied(self, value: bool):
+        self._data['stickied'] = value
+
+    @property
     def scheduled_at(self) -> datetime:
         """datetime: Scheduled time. This is used for scheduled events
         like maintenance in Cachet 2.3 were incident status is ``INCIDENT_SCHEDULED``.
@@ -121,6 +130,7 @@ class IncidentManager(Manager):
             message: str,
             status: int,
             visible: bool = True,
+            stickied: bool = False,
             component_id: int = None,
             component_status: int = None,
             notify: bool = True,
@@ -137,10 +147,12 @@ class IncidentManager(Manager):
             message (str): Mesage body for the issue
             status (int): Status of the incident (see enums)
             visible (bool): Publicly visible incident
+            stickied (bool): Stickied incident
             component_id (int): The component to update
             component_status (int): The status to apply on component
             notify (bool): If users should be notified
-            occurred_at: when the indicent was occurred
+            occurred_at: when the indicent occurred (cachet 2.4)
+            created_at: when the incident was created (cachet 2.3)
             template (str): Slug of template to use
             template_vars (list): Variables to the template
 
@@ -154,9 +166,11 @@ class IncidentManager(Manager):
                 'message': message,
                 'status': status,
                 'visible': 1 if visible else 0,
+                'stickied': 1 if stickied else 0,
                 'component_id': component_id,
                 'component_status': component_status,
                 'notify': 1 if notify else 0,
+                'created_at': created_at.strftime('%Y-%m-%d %H:%M') if created_at else None,
                 'occurred_at': occurred_at.strftime('%Y-%m-%d %H:%M') if occurred_at else None,
                 'template': template,
                 'vars': template_vars or [],
@@ -170,6 +184,7 @@ class IncidentManager(Manager):
             message: str = None,
             status: int = None,
             visible: bool = None,
+            stickied: bool = False,
             component_id: int = None,
             component_status: int = None,
             notify: bool = True,
@@ -188,10 +203,12 @@ class IncidentManager(Manager):
             message (str): Mesage body for the issue
             status (int): Status of the incident (see enums)
             visible (bool): Publicly visible incident
+            stickied (bool): Stickied incident
             component_id (int): The component to update
             component_status (int): The status to apply on component
             notify (bool): If users should be notified
             occurred_at: when the indicent was occurred
+            created_at: when the incident was created
             template (str): Slug of template to use
             template_vars (list): Variables to the template
 
@@ -209,9 +226,11 @@ class IncidentManager(Manager):
                 message=message,
                 status=status,
                 visible=1 if visible else 0,
+                stickied=1 if stickied else 0,
                 component_id=component_id,
                 component_status=component_status,
                 notify=1 if notify else 0,
+                created_at=created_at.strftime('%Y-%m-%d %H:%M') if created_at else None,
                 occurred_at=occurred_at.strftime('%Y-%m-%d %H:%M') if occurred_at else None,
                 template=template,
                 vars=template_vars,